+++ ChangeBrokerRestApi/DataUpdate.cs
@@ -1,0 +1,11 @@
+﻿using System.Text.Json;
+
+namespace ChangeBrokerRestApi
+{
+    public class DataUpdate
+    {
+        public string Bookmark { get; set; }
+        
+        public IEnumerable<Update> Data { get; set; }
+    }
+}