+++ ChangeBrokerRestApi/SignupCollection.cs
@@ -1,0 +1,7 @@
+﻿namespace ChangeBrokerRestApi
+{
+    public class SignupCollection
+    {
+        public IEnumerable<Signup> Signups { get; set; } = Enumerable.Empty<Signup>();
+    }
+}