+++ ChangeBrokerRestApi/SignupResource.cs
@@ -1,0 +1,9 @@
+﻿namespace ChangeBrokerRestApi
+{
+    public class SignupResource
+    {
+        public string ResourceId { get; set; }
+
+        public bool IsAccessible { get; set; }
+    }
+}