+++ ChangeBrokerRestApi/SignupResources.cs
@@ -1,0 +1,7 @@
+﻿namespace ChangeBrokerRestApi
+{
+    public class SignupResources
+    {
+        public IEnumerable<SignupResource> Resources { get; set; } = Enumerable.Empty<SignupResource>();
+    }
+}