+++ ChangeBrokerRestApi/SignupWithBookmark.cs
@@ -1,0 +1,8 @@
+﻿namespace ChangeBrokerRestApi
+{
+    [Serializable]
+    public class SignupWithBookmark : Signup
+    {
+        public string Bookmark { get; set; }
+    }
+}