+++ ChangeBrokerRestApi/Update.cs
@@ -1,9 +1,11 @@
-﻿namespace StreamingUpdatesRestApi
+﻿using System.Text.Json;
+
+namespace ChangeBrokerRestApi
 {
     public class Update
     {
         public string ResourceId { get; set; }
         public string Operation { get; set; }
-        public IEnumerable<SdsSimpleType> Events { get; set; }
+        public IEnumerable<JsonElement> Events { get; set; }
     }
 }