+++ StreamingUpdatesRestApi/DataUpdate.cs
@@ -1,7 +1,0 @@
-﻿namespace StreamingUpdatesRestApi
-{
-    public class DataUpdate
-    {
-        public IEnumerable<Update> Data { get; set; }
-    }
-}