+++ HISTORY.md
@@ -1,4 +1,21 @@
+# Version History
+
+## 1.0.4 / 2024-2-2
+
 - Migrate Pipelines
 
-## [' Hub StreamingUpdates Service ', 'NET REST API Sample\n'] / 2023-10-13
+## 1.0.3 / 2023-12-11
 
+- Change sample name from Streaming Updates to Change Broker
+
+## 1.0.2 / 2023-11-27
+
+- Update sample to include paging
+
+## 1.0.1 / 2023-11-7
+
+- Revise sample for API changes
+
+## 1.0.0 / 2023-10-2
+
+- Initial release